+++ src/App.js
@@ -1,14 +1,14 @@
 import React, { useState } from 'react';
 import { Global, css } from '@emotion/core';
 
-import Container from './components/Container';
-import Intro from './Intro';
-import Work from './Work';
-import History from './History';
-import Skill from './Skill';
-import Contact from './Contact';
-import Footer from './Footer';
-import Color from './components/Color';
+import Container from 'components/Container';
+import Intro from 'components/Intro/index';
+import Work from 'components/Work';
+import History from 'components/History';
+import Skill from 'components/Skill';
+import Contact from 'components/Contact';
+import Footer from 'components/Footer';
+import Color from 'components/Common/Color';
 import GitHubCorner from './components/Github-Corner';
 import { mq, theme } from './variables';
