+++ src/components/Common/Link.js
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "@emotion/styled";
 
-import { mq } from "../variables";
+import { mq } from "variables";
 
 const Link = styled.a`
   position: relative;
