+++ src/components/Common/Section.js
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "@emotion/styled";
 
-import { mq } from "../variables";
+import { mq } from "variables";
 
 const Section = styled.section`
   padding: 3rem 0;
